Move validators to utils package
