+++ pkg/artifactory/resource_artifactory_push_replication.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/utils"
 	"golang.org/x/exp/slices"
 )
 
@@ -58,7 +59,7 @@ var pushReplicationSchemaCommon = map[string]*schema.Schema{
 	"cron_exp": {
 		Type:             schema.TypeString,
 		Required:         true,
-		ValidateDiagFunc: validation.ToDiagFunc(validateCron),
+		ValidateDiagFunc: validation.ToDiagFunc(utils.ValidateCron),
 	},
 	"enable_event_replication": {
 		Type:     schema.TypeBool,
